feat(resonance): INTEGER resonance - no floats, quantum switchable
TOP-DOWN ARCHITECTURE (not bottom-up):
MahaKirtan controls mod_space (default 137), computes resonance directly.
No delegation to FluteSync/VinaSync. No translation layers.

FORMULA:
  flute_resonance = tick % mod_space  (integer, 0 to mod_space-1)
  vina_resonance = seed % mod_space   (integer, 0 to mod_space-1)

SWITCHABLE:
  mod_space=137 → quantum mode (MAHA_QUANTUM)
  mod_space=22  → shruti mode (SHRUTIS)

FILES CHANGED:
- maha_algorithm.py: KirtanComputeResult, MahaKirtanState, compute()
- _types.py: VibrationState, AkashState
- __init__.py: _akash init, _update_akash()
- steward.py: StewardResponse

REMOVED:
- shruti_raga.py (translation layer was wrong approach)

+++ vibe_core/mahamantra/cli/steward.py
@@ -111,8 +111,8 @@ class StewardResponse:
     chapter: int
     route: ResonanceRoute
     call_response: str  # "CALL" or "RESPONSE"
-    resonance: float  # Krishna's flute resonance (WHEN)
-    vina_resonance: float  # Narada's vina resonance (WHAT TYPE)
+    resonance: int  # Flute resonance = tick % mod_space (integer!)
+    vina_resonance: int  # Vina resonance = seed % mod_space (integer!)
     vina_string: int  # 1-5: CHAITANYA/NITYANANDA/ADVAITA/GADADHARA/SRIVASA
     shadow_phase: str  # "bhoga", "prasadam", or "return" (TRANSFORMATION)
     shadow_position: int  # 0-15 position in yajna cycle
